sigo-api/pyproject.toml
[project]
name = "sigo-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.0.1",
    "coverage>=7.12.0",
    "dotenv>=0.9.9",
    "fastapi>=0.121.3",
    "msal>=1.31.1",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.4",
    "python-jose[cryptography]>=3.3.0",
    "requests>=2.32.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
